build compile to compile stmts into bytecode

+++ include/runtime/value.h
@@ -3,7 +3,7 @@
 
 #include <stddef.h>
 
-typedef double Value;
+typedef int Value;
 
 struct ValueArray {
     Value *data;
